使用Code Inspect进行代码检查并改进结果

+++ main.cpp
@@ -1,4 +1,5 @@
 #include"generator.h"
+
 using namespace std;
 static int settle_flag = 0;
 
@@ -34,14 +35,14 @@ int check(int m, int n) {
 }
 
 
-void prune(int i, int j, bool point[10]){
+void prune(int i, int j, bool point[10]) {
     //行排除
-    for (int k = 0; k < 18; k+=2){
-        if (ques_board[i][k] != '0'&&k != j)point[ques_board[i][k] - '0'] = true;
+    for (int k = 0; k < 18; k += 2) {
+        if (ques_board[i][k] != '0' && k != j)point[ques_board[i][k] - '0'] = true;
     }
     //列排除
-    for (int k = 0; k < 9; k++){
-        if (ques_board[k][j] != '0'&&k != i)point[ques_board[k][j] - '0'] = true;
+    for (int k = 0; k < 9; k++) {
+        if (ques_board[k][j] != '0' && k != i)point[ques_board[k][j] - '0'] = true;
     }
     int m = 0, n = 0;
     if (i / 3 == 0)m = 0;
@@ -51,16 +52,16 @@ void prune(int i, int j, bool point[10]){
     else if (j / 6 == 1)n = 6;
     else if (j / 6 == 2)n = 12;
 
-    for (int c = m; c < m + 3; c++){
-        for (int d = n; d < n + 6; d+=2){
-            if(c!=i&&d!=j&&ques_board[c][d]!='0')point[ques_board[c][d] - '0'] = true;
+    for (int c = m; c < m + 3; c++) {
+        for (int d = n; d < n + 6; d += 2) {
+            if (c != i && d != j && ques_board[c][d] != '0')point[ques_board[c][d] - '0'] = true;
         }
     }
 }
 
 
-void settle(int pos){
-    if (pos == 162){
+void settle(int pos) {
+    if (pos == 162) {
         settle_flag = 1;
         return;
     }
@@ -68,56 +69,54 @@ void settle(int pos){
     i = pos / 18;
     j = pos % 18;
 
-    bool point[10] = { false };
+    bool point[10] = {false};
 
-    if (ques_board[i][j] == '0'){
+    if (ques_board[i][j] == '0') {
         prune(i, j, point);
-        for (k = 1; k <= 9; k++){
+        for (k = 1; k <= 9; k++) {
             if (point[k])
                 continue;
-            ques_board[i][j] = (char)('0' + k);
-            if (check(i, j)){
+            ques_board[i][j] = (char) ('0' + k);
+            if (check(i, j)) {
                 settle(pos + 2);
             }
             if (settle_flag) { return; }
             ques_board[i][j] = '0';//??
         }
-    }
-    else{
+    } else {
         settle(pos + 2);
     }
     if (settle_flag) { return; }
 }
 
 
-void settle_ques(){
-    int begin=0, end = 0;
-    int total=0,flag=0;
+void settle_ques() {
+    int begin = 0, end = 0;
+    int total = 0, flag = 0;
     FILE *fpQues, *fpSolution;
     char strSolution[200];
-    if(AbsolutePath[0]==0)
+    if (AbsolutePath[0] == 0)
         fpQues = fopen(QUESPATH, "r");
     else
         fpQues = fopen(AbsolutePath, "r");
     fpSolution = fopen(SUDOKUPATH, "w");
-    while (true)
-    {
-        for (int i = 0; i < 9; i++){
+    while (true) {
+        for (int i = 0; i < 9; i++) {
             fgets(ques_board[i], 20, fpQues);
         }
         flag = fgetc(fpQues);//读取中间的空行，判断是否文件尾
-        begin = clock();
+        begin = (int) clock();
         settle_flag = 0;
         settle(0);
-        end = clock();
+        end = (int) clock();
         total += end - begin;
         strSolution[0] = '\0';
-        for (int i = 0; i < 9; i++){
+        for (int i = 0; i < 9; i++) {
             strcat(strSolution, ques_board[i]);
         }
         if (flag == -1)
             strSolution[161] = '\0';
-        else{
+        else {
             strSolution[161] = '\n';
             strSolution[162] = '\n';
             strSolution[163] = '\0';
@@ -130,8 +129,7 @@ void settle_ques(){
 }
 
 
-
-int main(int argc, char** argv){
+int main(int argc, char **argv) {
 
     if (argc != 3)//参数不等于3个
     {
@@ -144,22 +142,19 @@ int main(int argc, char** argv){
         return 1;
     }
 
-    if ((strcmp(argv[1], "-c") != 0) && (strcmp(argv[1], "-s")!=0) && (strcmp(argv[1],"-p")!=0))
-    {
-        cout<< "Illegal parameter\n"
-            << "The first parameter should be -c or -s or -p\n"
-            << "-c means generating sudoku\n"
-            << "-s meas solve the problem read from the file\n";
+    if ((strcmp(argv[1], "-c") != 0) && (strcmp(argv[1], "-s") != 0) && (strcmp(argv[1], "-p") != 0)) {
+        cout << "Illegal parameter\n"
+             << "The first parameter should be -c or -s or -p\n"
+             << "-c means generating sudoku\n"
+             << "-s meas solve the problem read from the file\n";
         return 1;
     }
 
     if (!strcmp(argv[1], "-c"))//judge the second parameter when -c
     {
-        int len = strlen(argv[2]);
-        for (int i = 0; i < len; i++)
-        {
-            if (argv[2][i] > '9' || argv[2][i] < '0')
-            {
+        int len = (int) strlen(argv[2]);
+        for (int i = 0; i < len; i++) {
+            if (argv[2][i] > '9' || argv[2][i] < '0') {
                 cout << "Illegal parameter\n";
                 cout << "The second parameter should be a positive integer\n";
                 return 1;
@@ -167,8 +162,7 @@ int main(int argc, char** argv){
         }
 
         int number = 0;
-        for (int i = 0; i < len; i++)
-        {
+        for (int i = 0; i < len; i++) {
             number = number * 10 + argv[2][i] - '0';
         }
 
@@ -178,8 +172,7 @@ int main(int argc, char** argv){
         return 0;
     }
 
-    if (!strcmp(argv[1], "-s"))
-    {
+    if (!strcmp(argv[1], "-s")) {
         strcpy(AbsolutePath, argv[2]);
         settle_ques();
         cout << "The answer you need is in the sudoku.txt\n"
@@ -189,11 +182,9 @@ int main(int argc, char** argv){
 
     if (!strcmp(argv[1], "-p"))//default path ques.txt
     {
-        int len = strlen(argv[2]);
-        for (int i = 0; i < len; i++)
-        {
-            if (argv[2][i] > '9' || argv[2][i] < '0')
-            {
+        int len = (int) strlen(argv[2]);
+        for (int i = 0; i < len; i++) {
+            if (argv[2][i] > '9' || argv[2][i] < '0') {
                 cout << "Illegal parameter\n";
                 cout << "The second parameter should be a positive integer\n";
                 return 1;
@@ -201,8 +192,7 @@ int main(int argc, char** argv){
         }
 
         int number = 0;
-        for (int i = 0; i < len; i++)
-        {
+        for (int i = 0; i < len; i++) {
             number = number * 10 + argv[2][i] - '0';
         }
